requirements.txt
selenium
webdriver-manager
pytest
pytest-env
allure-pytest
openpyxl
